fix: new added signature available after reload only
