+++ src/views/settings/settings-view.tsx
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import React, { FC, useCallback, useMemo, useState } from 'react';
+import React, { FC, useCallback, useEffect, useMemo, useState } from 'react';
 
 import { Container, FormSection } from '@zextras/carbonio-design-system';
 import {
@@ -18,7 +18,7 @@ import {
 	useUserAccount,
 	useUserSettings
 } from '@zextras/carbonio-shell-ui';
-import { cloneDeep, filter, find, isEmpty, isEqual, map, reduce, remove } from 'lodash';
+import { cloneDeep, filter, find, isEmpty, isEqual, map, reduce, remove, unescape } from 'lodash';
 
 import { differenceIdentities, differenceObject, getPropsDiff } from './components/utils';
 import ComposeMessage from './compose-msg-settings';
@@ -33,7 +33,7 @@ import { TIMEOUTS } from '../../constants';
 import { NO_SIGNATURE_ID } from '../../helpers/signatures';
 import { useAppDispatch } from '../../hooks/redux';
 import { useUiUtilities } from '../../hooks/use-ui-utilities';
-import { SignatureRequest } from '../../store/actions/signatures';
+import { GetAllSignatures, SignatureRequest } from '../../store/actions/signatures';
 import type { AccountIdentity, PropsType, SignItemType } from '../../types';
 
 /* to keep track of changes done to props we use 3 different values:
@@ -72,7 +72,7 @@ const SettingsView: FC = () => {
 	const [currentIdentities, setCurrentIdentities] = useState(identities);
 	const [updatedIdentities, setUpdatedIdentities] = useState(identities);
 	const [signatures, setSignatures] = useState<SignItemType[]>(() => []);
-	const [originalSignatures, setOriginalSignatures] = useState(() => []);
+	const [originalSignatures, setOriginalSignatures] = useState<SignItemType[]>(() => []);
 	const [disabled, setDisabled] = useState(true);
 	const [flag, setFlag] = useState(false);
 
@@ -181,6 +181,40 @@ const SettingsView: FC = () => {
 		[]
 	);
 
+	const onSignaturesLoaded = useCallback(
+		(signs: Array<SignItemType>) => {
+			const signaturesItems = map(
+				signs,
+				(item: SignItemType, idx) =>
+					({
+						label: item.name,
+						name: item.name,
+						id: item.id,
+						description: unescape(item?.content?.[0]?._content)
+					}) as SignItemType
+			);
+			setSignatures(signaturesItems);
+			setOriginalSignatures(
+				signaturesItems.map((el) => ({
+					id: el.id,
+					name: el.label ?? '',
+					label: el.label ?? '',
+					description: el.description ?? ''
+				}))
+			);
+		},
+		[setOriginalSignatures, setSignatures]
+	);
+
+	// Fetches signatures from the BE
+	useEffect(() => {
+		GetAllSignatures()
+			.then(({ signature: signs }) => onSignaturesLoaded(signs))
+			.catch((err) => {
+				console.error(err);
+			});
+	}, [onSignaturesLoaded]);
+
 	const saveChanges = useCallback<SettingsHeaderProps['onSave']>(() => {
 		if (!isEqual(signatures, originalSignatures)) {
 			const validationResult = validateSignatures(signatures);
@@ -266,6 +300,11 @@ const SettingsView: FC = () => {
 					});
 					setFlag(!flag);
 					setDisabled(true);
+					GetAllSignatures()
+						.then(({ signature: signs }) => onSignaturesLoaded(signs))
+						.catch((err) => {
+							console.error(err);
+						});
 				} else {
 					createSnackbar({
 						key: `new`,
@@ -334,6 +373,7 @@ const SettingsView: FC = () => {
 		createSnackbar,
 		setNewOrForwardSignatureId,
 		flag,
+		onSignaturesLoaded,
 		updatedIdentities
 	]);
 
@@ -365,17 +405,12 @@ const SettingsView: FC = () => {
 					/>
 					<RecoverMessages />
 					<SignatureSettings
-						settingsObj={settingsObj}
 						setSignatures={setSignatures}
-						// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-						// @ts-ignore
 						setOriginalSignatures={setOriginalSignatures}
-						updateSettings={updateSettings}
 						updatedIdentities={updatedIdentities}
 						updateIdentities={updateIdentities}
 						setDisabled={setDisabled}
 						signatures={signatures}
-						flag={flag}
 					/>
 					{/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
 					{/* @ts-ignore */}
