+++ src/views/settings/signature-settings.tsx
@@ -18,14 +18,13 @@ import {
 	ButtonProps
 } from '@zextras/carbonio-design-system';
 import { t, useIntegratedComponent } from '@zextras/carbonio-shell-ui';
-import { map, unescape, reject, concat } from 'lodash';
+import { map, reject, concat } from 'lodash';
 import styled from 'styled-components';
 import { v4 as uuidv4 } from 'uuid';
 
 import SelectIdentitySignature from './components/select-identity-signature';
 import { signaturesSubSection, setDefaultSignaturesSubSection } from './subsections';
 import { NO_SIGNATURE_ID, NO_SIGNATURE_LABEL } from '../../helpers/signatures';
-import { GetAllSignatures } from '../../store/actions/signatures';
 import type { SignatureSettingsPropsType, SignItemType } from '../../types';
 
 const DeleteButton = styled(Button)`
@@ -82,40 +81,6 @@ const SignatureSettings: FC<SignatureSettingsPropsType> = ({
 		editorRef.current.editor = editor;
 	};
 
-	const onSignaturesLoaded = useCallback(
-		(signs: Array<SignItemType>) => {
-			const signaturesItems = map(
-				signs,
-				(item: SignItemType, idx) =>
-					({
-						label: item.name,
-						name: item.name,
-						id: item.id,
-						description: unescape(item?.content?.[0]?._content)
-					}) as SignItemType
-			);
-			setSignatures(signaturesItems);
-			setOriginalSignatures(
-				signaturesItems.map((el) => ({
-					id: el.id,
-					name: el.label ?? '',
-					label: el.label ?? '',
-					description: el.description ?? ''
-				}))
-			);
-		},
-		[setOriginalSignatures, setSignatures]
-	);
-
-	// Fetches signatures from the BE
-	useEffect(() => {
-		GetAllSignatures()
-			.then(({ signature: signs }) => onSignaturesLoaded(signs))
-			.catch((err) => {
-				console.error(err);
-			});
-	}, [onSignaturesLoaded]);
-
 	// Set the default current signature if missing
 	useEffect(() => {
 		if (signatures?.length && !currentSignature) {
@@ -126,7 +91,7 @@ const SignatureSettings: FC<SignatureSettingsPropsType> = ({
 	// Creates an empty signature
 	const createEmptySignature = useCallback(
 		(): SignItemType => ({
-			id: uuidv4(),
+			id: `unsaved-signature-${uuidv4()}`,
 			label: t('label.enter_name', 'Enter Name'),
 			name: t('label.enter_name', 'Enter Name'),
 			description: ''
@@ -157,10 +122,12 @@ const SignatureSettings: FC<SignatureSettingsPropsType> = ({
 	// Composes the SelectItem array for the signature selects
 	const signatureSelectItems: SelectItem[] = useMemo(
 		(): SelectItem[] =>
-			concat(noSignature, signatures).map((signature) => ({
-				label: signature.label,
-				value: signature.id
-			})),
+			concat(noSignature, signatures)
+				.filter((signature) => !signature.id.includes('unsaved-signature-'))
+				.map((signature) => ({
+					label: signature.label,
+					value: signature.id
+				})),
 		[noSignature, signatures]
 	);
